Fixed follow-up SAM3 crash ("Got unsupported ScalarType BFloat16") that surfaced on the same Torch/CUDA combinations once the autocast wrapper was added: under torch.autocast(bfloat16) the processor's masks and scores tensors come back as bf16, and calling .cpu().numpy() on them fails because NumPy has no bfloat16 dtype. backend.py SAM3Engine.generate_mask now calls .float() before .cpu().numpy() on both tensors to promote them to float32 first. Negligible overhead since masks/scores are tiny compared to the model itself, and it works uniformly whether autocast leaves them as bf16 or fp32.

+++ backend.py
@@ -760,8 +760,9 @@ class SAM3Engine:
             
             if masks is None: return None, "No masks returned"
             
-            if hasattr(masks, "cpu"): masks = masks.cpu().numpy()
-            if hasattr(scores, "cpu"): scores = scores.cpu().numpy()
+            # .float() before .numpy() — autocast can leave these as bfloat16, which numpy doesn't support
+            if hasattr(masks, "cpu"): masks = masks.float().cpu().numpy()
+            if hasattr(scores, "cpu"): scores = scores.float().cpu().numpy()
             
             if masks.size == 0: return None, "No masks found"
